+++ steel/src/primitives/hashmaps.rs
@@ -92,7 +92,7 @@ impl HashMapOperations {
             if let SteelVal::HashMapV(hm) = hashmap {
                 match hm.get(key) {
                     Some(v) => Ok(v.clone()),
-                    None => Ok(SteelVal::BoolV(true)),
+                    None => Ok(SteelVal::BoolV(false)),
                 }
             } else {
                 stop!(TypeMismatch => "hm-insert takes a hashmap")
